pyproject.toml
[project]
name = "graph-bench"
version = "0.1.0"
description = "Comprehensive benchmark suite for graph databases"
readme = "README.md"
requires-python = ">=3.12"
dependencies = []
authors = [{ name = "Steven Grond" }]
license = { text = "Apache-2.0" }
keywords = [
    "benchmark",
    "graph-database",
    "neo4j",
    "memgraph",
    "ladybug",
    "arangodb",
    "duckdb",
    "falkordb",
    "nebula",
    "grafeo",
    "performance",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Database",
    "Topic :: Software Development :: Testing",
]

[project.urls]
Repository = "https://github.com/StevenBtw/graph-bench"

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.14",
    "prek>=0.3.1",
    "ty>=0.0.14",
]
cli = ["typer>=0.15.0"]
neo4j = ["neo4j>=5.0"]
memgraph = ["neo4j>=5.0"]
ladybug = ["real_ladybug>=0.14"]
arangodb = ["python-arango>=7.0"]
grafeo = ["grafeo>=0.2.0"]
duckdb = ["duckdb>=1.0"]
falkordb = ["FalkorDB>=1.0"]
nebula = ["nebula3-python>=3.0"]
datasets = ["faker>=20.0"]
reporting = ["matplotlib>=3.8", "pandas>=2.0"]
all-adapters = [
    "neo4j>=5.0",
    "real_ladybug>=0.14",
    "python-arango>=7.0",
    "grafeo>=0.2.0",
    "duckdb>=1.0",
    "FalkorDB>=1.0",
    "nebula3-python>=3.0",
]
all = [
    "graph-bench[dev,cli,all-adapters,datasets,reporting]",
]

[project.scripts]
graph-bench = "graph_bench.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["graph_bench"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
addopts = "--cov=graph_bench --cov-report=term-missing --cov-fail-under=88"
testpaths = ["tests"]
markers = [
    "slow: slow tests (run with --runslow)",
    "integration: integration tests requiring database",
    "neo4j: requires Neo4j",
    "memgraph: requires Memgraph",
    "kuzu: requires Kuzu",
    "arangodb: requires ArangoDB",
    "grafeo: requires Grafeo",
    "duckdb: requires DuckDB",
]

[tool.coverage.run]
source = ["graph_bench"]
omit = ["*/__init__.py"]

[tool.coverage.report]
fail_under = 88
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[tool.ty.environment]
python-version = "3.12"
